refactor(condo): DOMA-7032 refactored ChangeHistory
