+++ apps/condo/domains/common/components/ChangeHistory/ChangeHistoryDiff.tsx
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export const ChangeHistoryDiff: React.FC = (props) => <p {...props} />
