+++ apps/condo/pages/ticket/[id]/index.tsx
@@ -23,6 +23,7 @@ import { useOrganization } from '@open-condo/next/organization'
 import { ActionBar, Alert, Button } from '@open-condo/ui'
 
 import { ChangeHistory } from '@condo/domains/common/components/ChangeHistory'
+import { HistoricalChange } from '@condo/domains/common/components/ChangeHistory/HistoricalChange'
 import { Comments } from '@condo/domains/common/components/Comments'
 import { AccessDeniedPage } from '@condo/domains/common/components/containers/AccessDeniedPage'
 import { PageContent, PageWrapper } from '@condo/domains/common/components/containers/BaseLayout'
@@ -629,6 +630,7 @@ export const TicketPageContent = ({ ticket, refetchTicket, loading, organization
                         title={TicketChangesMessage}
                         useChangedFieldMessagesOf={useTicketChangedFieldMessagesOf}
                         Diff={TicketChangeDiff}
+                        HistoricalChange={HistoricalChange}
                     />
                     <Col span={24}>
                         <TicketActionBar
